+++ dll/echarts/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.0.6](https://git.easyops.local/anyclouds/next-core/compare/@dll/echarts@1.0.5...@dll/echarts@1.0.6) (2020-08-26)
+
+**Note:** Version bump only for package @dll/echarts
+
 ## [1.0.5](https://git.easyops.local/anyclouds/next-core/compare/@dll/echarts@1.0.4...@dll/echarts@1.0.5) (2020-06-05)
 
 **Note:** Version bump only for package @dll/echarts